Split unit test in tasks.rs
